proj/Smart_Laundry/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-countdowntimer-model==0.0.7
django-extensions==3.1.1
django-rest-framework==0.1.0
djangorestframework==3.12.2
mysqlclient==2.0.3
pytz==2020.5
sqlparse==0.4.1
